requirements.txt
google-generativeai>=0.7.0
streamlit>=1.35.0
requests
beautifulsoup4
pypdf
